Fixed crash caused by array overflow in summary.c
This patch fixes a crash that is caused by an overflow of the fname
array in summary.c. fname was originally defined to be 256 bytes, however
name_root, which is defined to be 256 bytes, and additional arguments
are written to fname via an sprintf call. This can potentially overflow
the amount of memory that was allocated for the fname array and result
in a crash.

Signed-off-by: jmather-sesi <54645798+jmather-sesi@users.noreply.github.com>